...G...... [DEV-681] Fixed Windows agent compilation warnings
git-svn-id: svn://svn.zabbix.com/branches/dev@74058 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ DEV-681/include/comms.h
@@ -29,8 +29,8 @@ typedef long	ssize_t;
 #endif
 
 #ifdef _WINDOWS
-#	define ZBX_TCP_WRITE(s, b, bl)	((ssize_t)send((s), (b), (bl), 0))
-#	define ZBX_TCP_READ(s, b, bl)	((ssize_t)recv((s), (b), (bl), 0))
+#	define ZBX_TCP_WRITE(s, b, bl)	((ssize_t)send((s), (b), (int)(bl), 0))
+#	define ZBX_TCP_READ(s, b, bl)	((ssize_t)recv((s), (b), (int)(bl), 0))
 #	define zbx_socket_close(s)	if (ZBX_SOCKET_ERROR != (s)) closesocket(s)
 #	define zbx_socket_last_error()	WSAGetLastError()
